+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Added
+- Clicking the HoverPreview will now play the video
+
 ## Changed
 - Set bounds on options for HoverPreview and SeekPreview